mlops-unit1/requirements.txt
absl-py==2.3.1
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
altair==6.0.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
arrow==1.4.0
astunparse==1.6.3
attrs==25.4.0
autopep8==2.3.2
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.4
bitsandbytes==0.48.2
black==25.9.0
blinker==1.9.0
build==1.4.0
cachetools==6.2.6
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
chromadb==1.5.0
click==8.2.1
codecarbon==3.0.8
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cssselect==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.4.0
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
durationpy==0.10
et_xmlfile==2.0.0
feedfinder2==0.0.4
feedparser==6.0.11
fief-client==0.20.0
filelock==3.18.0
Flask==3.0.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.7.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
googleapis-common-protos==1.72.0
gotrue==1.3.1
greenlet==3.3.1
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.33.4
idna==3.10
imbalanced-learn==0.14.1
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jwcrypto==1.5.6
keras==3.12.0
kiwisolver==1.4.9
kubernetes==35.0.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.11
langchain-text-splitters==1.1.0
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.5
langsmith==0.7.1
libclang==18.1.1
lxml==6.0.0
lxml_html_clean==0.4.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.16.0
networkx==3.5
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.6
nvidia-ml-py==13.580.82
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.24.1
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
postgrest==0.13.2
posthog==5.4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pybase64==1.4.3
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.7.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.2
questionary==2.1.1
RapidFuzz==3.14.3
realtime==1.0.6
referencing==0.37.0
regex==2024.11.6
requests==2.32.5
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
sentence-transformers==5.0.0
sgmllib3k==1.0.0
shellingham==1.5.4
simsimd==6.5.0
six==1.17.0
sklearn-compat==0.1.5
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.46
storage3==0.6.1
streamlit==1.54.0
StrEnum==0.4.15
stringzilla==3.12.5
supabase==2.0.0
supafunc==0.3.3
sympy==1.14.0
ta==0.11.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.53.3
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uuid_utils==0.14.0
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==12.0
Werkzeug==3.1.4
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
yaspin==3.3.0
zipp==3.23.0
zstandard==0.25.0
